+++ training-app/config/routes.rb
@@ -1,7 +1,13 @@
+# frozen_string_literal: true
+
 Rails.application.routes.draw do
+  # User authentication
   mount_devise_token_auth_for 'User', at: 'auth'
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  root to: 'home#index'
-  # Add only index and show routes
-  resources :books, only: %I[index show]
+  # API entries are inside a namespace
+  namespace :api do
+    # First (and only) version of the API
+    namespace :v1 do
+      resources :books, only: %I[index show]
+    end
+  end
 end
